+++ src/main/java/net/oneandone/sushi/fs/http/HttpNode.java
@@ -32,7 +32,7 @@ import net.oneandone.sushi.fs.Node;
 import net.oneandone.sushi.fs.NodeNotFoundException;
 import net.oneandone.sushi.fs.SetLastModifiedException;
 import net.oneandone.sushi.fs.SizeException;
-import net.oneandone.sushi.fs.http.methods.GenericMethod;
+import net.oneandone.sushi.fs.http.model.GenericMethod;
 import net.oneandone.sushi.fs.http.model.Body;
 import net.oneandone.sushi.fs.http.model.Header;
 import net.oneandone.sushi.fs.http.model.ProtocolException;