+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 3.1.3 (pending)
 
-* added HttpNode.put method
+* improved REST support
+  * added HttpNode.put method (and changed the existing code to also consider status code 200 as a success response)
 
 
 ### 3.1.2 (2016-07-26)